Refactored display Items function separating searchArtistAlbums into a separated module
