+++ src/index.js
@@ -4,20 +4,13 @@ import Comment from './modules/comment.js';
 import initId from './modules/init.js';
 import cardCounter from './modules/cardCounter.js';
 import updateAllLikes from './modules/likesData.js';
-import searchArtistId from './modules/searchArtist.js';
+import searchArtistId from './modules/searchArtistId.js';
+import searchArtistAlbums from './modules/searchArtistAlbums.js';
+
 // Display all items
 const displayItems = async (artistId = '271256') => {
   const container = document.getElementById('section');
-  const response = await fetch(
-    `https://itunes.apple.com/lookup?id=${artistId}&entity=album&limit=12`,
-    {
-      method: 'POST',
-      headers: {
-        'Content-type': 'application/json; charset=UTF-8',
-      },
-    },
-  );
-  const obj = await response.json();
+  const obj = await searchArtistAlbums(artistId, 12);
   for (let index = 1; index < obj.results.length; index += 1) {
     const element = obj.results[index];
     displayCard(
